+++ axisubs/app/Models/Site/Plans.php
@@ -158,8 +158,10 @@ class Plans extends Post{
         if($id){
             $postDB = Post::where('post_type', 'axisubs_plans')->get();
             $postTable = $postDB->find($id);
-            $postTable->meta()->delete();
-            return $postTable->delete();
+            if(!empty($postTable)){
+                $postTable->meta()->delete();
+                return $postTable->delete();
+            }
         } else {
             return false;
         }
